Refactor interface to SaveCancel.
The reason the couplings seemed so awkward was the assumption
that SaveCancel had to know what to return. Quite wrong. It
just has to return True/False according to whether the save
button was the trigger. Its owner then decides what (if any)
action to take.

Still to be addressed:

a) Non-appearance of final element in widget layout.
b) Inability to center button row (suspect a simple
   styling change might fix it).