Simplify as much as possible structurally.
Key insight for componentisation was to simplify the structures
as much as possible and keep communications clear. The SaveCancel
still has the issue that it's not reusable, because it uses
fixed ids, but it will do for now, and it's possible that textual
isn't as picky as HTML about unique ids.

A better first approach might be to switch from ids to classes,
which don't have the same uniqueness requirements. Anyway, for
now this approach seems satifactory.